Rearranged directories, and moved errors to be called from pkg
