+++ internal/errorutil/analyze.go
