+++ internal/errorutil/constants.go
