+++ internal/errorutil/types.go
