+++ internal/errorutil/walk.go
